move cron job to run.js file and initialize global mean and variance values
