+++ backend/routes/statDictionaries.js
@@ -1,0 +1,53 @@
+let STAT_MEANS = {
+  minutes_played: 0,
+  field_goals,
+  field_attempts: 0,
+  field_percent: 0.0,
+  three_fg: 0,
+  three_attempts: 0,
+  three_percent: 0,
+  two_fg: 0,
+  two_attempts: 0,
+  two_percent: 0,
+  effect_fg_percent: 0,
+  ft: 0,
+  fta: 0,
+  ft_percent: 0,
+  ORB: 0,
+  DRB: 0,
+  TRB: 0,
+  AST: 0,
+  STL: 0,
+  BLK: 0,
+  TOV: 0,
+  PF: 0,
+  PTS: 0,
+};
+
+let STAT_VARIANCES = {
+  minutes_played: 0,
+  field_goals,
+  field_attempts: 0,
+  field_percent: 0.0,
+  three_fg: 0,
+  three_attempts: 0,
+  three_percent: 0,
+  two_fg: 0,
+  two_attempts: 0,
+  two_percent: 0,
+  effect_fg_percent: 0,
+  ft: 0,
+  fta: 0,
+  ft_percent: 0,
+  ORB: 0,
+  DRB: 0,
+  TRB: 0,
+  AST: 0,
+  STL: 0,
+  BLK: 0,
+  TOV: 0,
+  PF: 0,
+  PTS: 0,
+};
+
+export { STAT_MEANS, STAT_VARIANCES };
